fix(admin-api): classify and retry accept errors instead of terminating
accept_loop propagated any listener.accept() failure via ?, so one transient
ECONNABORTED or a momentary EMFILE ended the admin server -- and an embedder
racing RunningServer::wait() turns that into a node leaving the cluster, making
fd pressure a fleet-wide correlated restart. Apply the data plane's #750 policy:
transient errors retry immediately, systemic ones back off and log once per
outage. A broken listener (EBADF/ENOTSOCK/EINVAL) still terminates so wait()
keeps reporting a genuinely dead admin plane. The shared machinery moves from
imposter::manager to proxy::network and becomes public so the two listeners
cannot drift apart.
