+++ crates/rift-mock-core/src/imposter/manager.rs
@@ -12,6 +12,9 @@ use crate::behaviors::ResponseSequencer;
 use crate::extensions::decorate::ResponseDecorator;
 use crate::extensions::flow_state::FlowStoreProvider;
 use crate::imposter::journal::RequestJournal;
+use crate::proxy::network::{
+    AcceptBackoff, AcceptErrorClass, AcceptErrorLog, classify_accept_error,
+};
 use crate::recording::ProxyRecordingStore;
 use arc_swap::ArcSwapOption;
 use hyper::service::service_fn;
@@ -24,100 +27,6 @@ use std::time::Duration;
 use tokio::net::TcpListener;
 use tokio::sync::broadcast;
 
-// ── Accept-error handling (issue #750) ───────────────────────────────────────────────────────
-//
-// A bare `accept(2)` error arm that logs and immediately retries has two coupled defects, both
-// amplified ×N under the per-core listener fan-out (#745): it spins hot on a *systemic* error
-// (fd exhaustion cannot be cured by retrying), and it logs per failed accept — the exact
-// per-event-rate logging trap the journal-cap fix (#741) removed. The pieces below fix both with
-// no cost on the happy path.
-
-/// How an `accept(2)` error should be handled.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AcceptErrorClass {
-    /// Expected under load; the next accept is likely fine — retry immediately, log at `debug`.
-    Transient,
-    /// Resource exhaustion (or an unknown error): retrying now cannot succeed until something
-    /// frees, so back off. Unknown errors land here deliberately — spinning on an unrecognized
-    /// failure is the worse outcome.
-    Systemic,
-}
-
-/// Classify via [`std::io::ErrorKind`] so no `libc` dependency is needed: only the transient
-/// classes are named (`ECONNABORTED`/`EINTR`/`ECONNRESET`); everything else — resource exhaustion
-/// like `EMFILE`/`ENFILE`/`ENOBUFS` (`Uncategorized`) or `ENOMEM` (`OutOfMemory`), and any truly
-/// unknown error — falls through `_` to systemic.
-fn classify_accept_error(e: &std::io::Error) -> AcceptErrorClass {
-    use std::io::ErrorKind::{ConnectionAborted, ConnectionReset, Interrupted};
-    match e.kind() {
-        ConnectionAborted | Interrupted | ConnectionReset => AcceptErrorClass::Transient,
-        _ => AcceptErrorClass::Systemic,
-    }
-}
-
-/// Exponential backoff for the systemic accept-error path: 1ms doubling to a 1s cap, reset on the
-/// first successful accept. Pure and clock-free so the schedule is unit-testable without sockets.
-struct AcceptBackoff {
-    current: Duration,
-}
-
-impl AcceptBackoff {
-    const INITIAL: Duration = Duration::from_millis(1);
-    const MAX: Duration = Duration::from_secs(1);
-
-    fn new() -> Self {
-        Self {
-            current: Self::INITIAL,
-        }
-    }
-
-    /// The delay to wait now; advances the schedule (double, capped) for the next call.
-    fn next_delay(&mut self) -> Duration {
-        let delay = self.current;
-        self.current = (self.current * 2).min(Self::MAX);
-        delay
-    }
-
-    fn reset(&mut self) {
-        self.current = Self::INITIAL;
-    }
-}
-
-/// Once-per-transition logging for the systemic accept-error path, mirroring the journal-cap warn
-/// fix (#741): one `error!` on entering the error state, silence while in it (counting suppressed
-/// errors), one `info!` carrying the suppressed count on recovery.
-#[derive(Default)]
-struct AcceptErrorLog {
-    in_error: bool,
-    suppressed: u64,
-}
-
-impl AcceptErrorLog {
-    /// Record a systemic error. Returns `true` **only** on the transition *into* the error state
-    /// (emit the single `error!`); later errors return `false` and bump the suppressed count.
-    fn on_error(&mut self) -> bool {
-        if self.in_error {
-            self.suppressed += 1;
-            false
-        } else {
-            self.in_error = true;
-            self.suppressed = 0;
-            true
-        }
-    }
-
-    /// Record a successful accept. Returns `Some(suppressed)` **only** on the transition *out* of
-    /// the error state (emit the single recovery `info!`); `None` in steady state, so a healthy
-    /// accept path pays just one branch.
-    fn on_success(&mut self) -> Option<u64> {
-        if self.in_error {
-            self.in_error = false;
-            Some(std::mem::take(&mut self.suppressed))
-        } else {
-            None
-        }
-    }
-}
 use tracing::{debug, error, info, warn};
 
 /// Server-level TLS defaults for HTTPS imposters that don't carry their own cert/key (issue #206).
@@ -1629,92 +1538,6 @@ mod tests {
     fn raise_fd_limit() {}
 
     // ── Accept-error handling (issue #750) ───────────────────────────────────────────────────
-    mod accept_error {
-        use super::*;
-        use std::io::{Error, ErrorKind};
-
-        #[test]
-        fn classify_transient_vs_systemic() {
-            for kind in [
-                ErrorKind::ConnectionAborted,
-                ErrorKind::Interrupted,
-                ErrorKind::ConnectionReset,
-            ] {
-                assert_eq!(
-                    classify_accept_error(&Error::from(kind)),
-                    AcceptErrorClass::Transient,
-                    "{kind:?} must be transient (retry immediately)"
-                );
-            }
-            // EMFILE (24) / ENFILE (23) map to Uncategorized -> systemic; and a named non-transient
-            // kind is systemic too. Unknown must never be treated as transient.
-            for e in [
-                Error::from_raw_os_error(24), // EMFILE
-                Error::from_raw_os_error(23), // ENFILE
-                Error::from(ErrorKind::OutOfMemory),
-                Error::from(ErrorKind::Other),
-            ] {
-                assert_eq!(
-                    classify_accept_error(&e),
-                    AcceptErrorClass::Systemic,
-                    "{e:?} must back off, not spin"
-                );
-            }
-        }
-
-        #[test]
-        fn backoff_doubles_from_1ms_caps_at_1s_and_resets() {
-            let mut b = AcceptBackoff::new();
-            assert_eq!(
-                b.next_delay(),
-                Duration::from_millis(1),
-                "first delay is 1ms, not 0"
-            );
-            assert_eq!(b.next_delay(), Duration::from_millis(2));
-            assert_eq!(b.next_delay(), Duration::from_millis(4));
-            assert_eq!(b.next_delay(), Duration::from_millis(8));
-            // Fast-forward well past the cap.
-            for _ in 0..20 {
-                b.next_delay();
-            }
-            assert_eq!(b.next_delay(), Duration::from_secs(1), "capped at 1s");
-            b.reset();
-            assert_eq!(
-                b.next_delay(),
-                Duration::from_millis(1),
-                "reset returns to 1ms"
-            );
-        }
-
-        // The log state machine backs the "exactly one error!, exactly one info! with the
-        // suppressed count" AC without sockets: on_error transitions once, on_success reports the
-        // count once, steady state stays silent.
-        #[test]
-        fn log_emits_once_per_transition_with_suppressed_count() {
-            let mut log = AcceptErrorLog::default();
-            assert!(
-                log.on_error(),
-                "first systemic error -> emit the one error!"
-            );
-            assert!(!log.on_error(), "second is suppressed");
-            assert!(!log.on_error(), "third is suppressed");
-            assert_eq!(
-                log.on_success(),
-                Some(2),
-                "recovery emits one info! carrying the 2 suppressed errors"
-            );
-            assert_eq!(log.on_success(), None, "steady state is silent");
-            assert_eq!(log.on_success(), None);
-
-            // A fresh outage re-arms: one error!, count restarts from 0.
-            assert!(log.on_error());
-            assert_eq!(
-                log.on_success(),
-                Some(0),
-                "immediate recovery suppressed nothing"
-            );
-        }
-    }
 
     #[tokio::test]
     async fn test_create_imposter_writes_to_datadir() {
